requirements.txt
striprtf>=0.0.26
python-docx>=1.1.0
